requirements.txt
streamlit>=1.20.0
opencv-python-headless
mediapipe
pandas
numpy
python-pptx
Pillow
